Implement steps for plays
